+++ lib/moco.rb
@@ -6,6 +6,11 @@ require "active_support/inflector"
 
 require_relative "moco/version"
 
+# Core classes needed by entities
+require_relative "moco/connection"
+require_relative "moco/collection_proxy"
+require_relative "moco/nested_collection_proxy"
+
 # New API (v2)
 require_relative "moco/entities/base_entity"
 require_relative "moco/entities/project"
@@ -22,9 +27,6 @@ require_relative "moco/entities/presence"
 require_relative "moco/entities/holiday"
 require_relative "moco/entities/planning_entry"
 require_relative "moco/client"
-require_relative "moco/connection"
-require_relative "moco/collection_proxy"
-require_relative "moco/nested_collection_proxy"
 require_relative "moco/entity_collection"
 
 require_relative "moco/sync"